Fix tests, modify cu
