--- .github/workflows/interchain-test.yml
@@ -1,95 +1,0 @@
-name: interchain-test
-permissions:
-  contents: read
-on:
-  push:
-    branches: [main]
-    paths:
-      - interchain/**
-      - framework/**
-      - modules/**
-  pull_request:
-    paths:
-      - interchain/**
-      - framework/**
-      - modules/**
-
-env:
-  SCCACHE_GHA_ENABLED: "true"
-  RUSTC_WRAPPER: "sccache"
-
-jobs:
-  framework-clone-tests:
-    runs-on: ubuntu-latest
-    name: framework clone tests
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          submodules: true
-      - uses: webfactory/ssh-agent@v0.8.0
-        with:
-          ssh-private-key: |
-            ${{ secrets.SSH_PRIVATE_KEY_MULTI_TEST }}
-            ${{ secrets.SSH_PRIVATE_KEY_CW_ORCH_INTERCHAIN }}
-      - name: Setup sccache
-        uses: mozilla-actions/sccache-action@v0.0.5
-        with:
-          version: "v0.8.1"
-      - name: Install stable
-        uses: dtolnay/rust-toolchain@master
-        with:
-          toolchain: stable
-      - name: cargo generate-lockfile
-        if: hashFiles('Cargo.lock') == ''
-        working-directory: ./interchain/framework-clone-testing
-        run: cargo generate-lockfile
-      - uses: taiki-e/install-action@nextest
-      - name: save private network configs
-        run: |
-          mkdir -p ~/.cw-orchestrator
-          echo -n '${{ secrets.CW_ORCH_NETWORKS }}' > ~/.cw-orchestrator/networks.toml
-      - uses: Swatinem/rust-cache@v2
-        with:
-          workspaces: "interchain/framework-clone-testing -> target"
-      - name: cargo test --locked
-        env:
-          SCCACHE_GHA_ENABLED: "true"
-          RUSTC_WRAPPER: "sccache"
-        working-directory: ./interchain/framework-clone-testing
-        run: cargo nextest run --retries 3
-
-  modules-clone-tests:
-    runs-on: ubuntu-latest
-    name: module clone tests
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          submodules: true
-      - uses: webfactory/ssh-agent@v0.8.0
-        with:
-          ssh-private-key: |
-            ${{ secrets.SSH_PRIVATE_KEY_MULTI_TEST }}
-            ${{ secrets.SSH_PRIVATE_KEY_CW_ORCH_INTERCHAIN }}
-      - name: Setup sccache
-        uses: mozilla-actions/sccache-action@v0.0.5
-        with:
-          version: "v0.8.1"
-      - name: Install stable
-        uses: dtolnay/rust-toolchain@master
-        with:
-          toolchain: stable
-      - name: cargo generate-lockfile
-        if: hashFiles('Cargo.lock') == ''
-        working-directory: ./interchain/modules-clone-testing
-        run: cargo generate-lockfile
-      - uses: taiki-e/install-action@nextest
-      - name: save private network configs
-        run: |
-          mkdir -p ~/.cw-orchestrator
-          echo -n '${{ secrets.CW_ORCH_NETWORKS }}' > ~/.cw-orchestrator/networks.toml
-      - uses: Swatinem/rust-cache@v2
-        with:
-          workspaces: "interchain/modules-clone-testing -> target"
-      - name: cargo test --locked
-        working-directory: ./interchain/modules-clone-testing
-        run: cargo nextest run --retries 3
