+++ framework/artifacts/checksums.txt
@@ -1,8 +1,8 @@
-a2c127a1f275484c16059bccab2c50dca6ba6c8753b947bdf48870d3ab64af60  abstract_account-xion.wasm
-4e6c8d287d806ea97288cd78e5898c727c9e90a82e423aaeed2d824617b3e83b  abstract_account.wasm
-706f0d8b007c0752c2f795e4a11d6fc42da9ff36485c6efcf2abd32c541d1fa9  abstract_ans_host.wasm
-6d2ff01481324c369188d052be01182dcb6633c2623e9d855465b8b4804caf31  abstract_ibc_client.wasm
-18f6be7e4a87f9083606b3ae7040ef07df363096b66e1fb780463856afcbd98e  abstract_ibc_host.wasm
-8561e3c0552cca824fdd3859e1771a9594c41d5824bcd064749b91b89e92d3b1  abstract_ica_client.wasm
-5b250e3d2820b24558f2ccda57496b70ea4223f74e30fd4d68e21ebe74b5a3c8  abstract_module_factory.wasm
-1bb37f63acd50dfcdb280c2f1dc49e94a2cb3402d272e1620ae3c00e3a605deb  abstract_registry.wasm
+0ff360e4f928b33457ddd8e5680368ce6af2ade17a18f697bd0e52870d1ae771  abstract_account-xion.wasm
+26b3cd6c32e0adb9597e755e5cc1a5bb5456a116b896241d3dae411092a230af  abstract_account.wasm
+bceb147b9c6cc5ef5a9372d9ec08c0461d54bc417995fc940e2d8f89610239f3  abstract_ans_host.wasm
+0585dd63fe993f34e42abf3cbab5162f06fdd66bf8d93789b389e09dcc7b7d1a  abstract_ibc_client.wasm
+37f2778d592ff561384b8b79f433dcb484b2875b204529bd243860e0eb058edb  abstract_ibc_host.wasm
+e950929fa76e4eed86ff0afb6c1f3b72971bdd683250943caacff9ad51ed3a5f  abstract_ica_client.wasm
+e89369e01a70d001ab8a2177dc51ba69a85c8163404e0904735de8ebdef9ce98  abstract_module_factory.wasm
+f1cf814a0468311512a30dd0499ddba69d64df093ce069cfa4fbfaf66899798a  abstract_registry.wasm
