--- framework/packages/abstract-interface/tests/integration_tests.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "interchain")]
-mod interchain_integration;
