+++ framework/tests/src/interchain/mod.rs
@@ -5,13 +5,14 @@ use abstract_std::objects::{AccountId, AccountTrace, TruncatedChainId};
 use anyhow::Result as AnyResult;
 use cosmwasm_std::{coins, Coin};
 use cw_orch::{
+    anyhow,
     daemon::{TxSender, Wallet},
     prelude::*,
 };
 use cw_orch_interchain::{
     core::{IbcQueryHandler, InterchainEnv},
     daemon::DaemonInterchain,
-    prelude::Starship,
+    prelude::*,
 };
 
 pub const JUNO: &str = "juno-1";
@@ -120,4 +121,29 @@ pub fn abstract_starship_interfaces(
     )?;
 
     Ok((abstr_juno, abstr_juno2))
+}
+
+pub fn ibc_abstract_setup<Chain: IbcQueryHandler<Sender = Addr>, IBC: InterchainEnv<Chain>>(
+    interchain: &IBC,
+    origin_chain_id: &str,
+    remote_chain_id: &str,
+) -> anyhow::Result<(Abstract<Chain>, Abstract<Chain>)> {
+    let mut origin_chain = interchain.get_chain(origin_chain_id).unwrap();
+    let mut remote_chain = interchain.get_chain(remote_chain_id).unwrap();
+
+    origin_chain.set_sender(Abstract::mock_admin(&origin_chain));
+    remote_chain.set_sender(Abstract::mock_admin(&remote_chain));
+    // Deploying abstract and the IBC abstract logic
+    let abstr_origin = Abstract::deploy_on_mock(origin_chain.clone())?;
+    let abstr_remote = Abstract::deploy_on_mock(remote_chain.clone())?;
+
+    abstr_origin.connect_to(&abstr_remote, interchain)?;
+
+    Ok((abstr_origin, abstr_remote))
+}
+
+/// This allows env_logger to start properly for tests
+/// The logs will be printed only if the test fails !
+pub fn logger_test_init() {
+    let _ = env_logger::builder().is_test(true).try_init();
 }
