+++ framework/tests/tests/clone_tests/mod.rs
@@ -1,0 +1,2 @@
+mod migrate;
+mod upgrade_account;
