+++ framework/tests/tests/interchain/mod.rs
@@ -1,0 +1,1 @@
+mod interchain_accounts;
