+++ framework/tests/tests/mod.rs
@@ -1,0 +1,3 @@
+#[cfg(test)]
+mod clone_tests;
+mod interchain;
